+++ assets/snippets/DocLister/core/controller/site_content.php
@@ -456,7 +456,7 @@ class site_contentDocLister extends DocLister
         $sanitarInIDs = $this->sanitarIn($this->IDs);
 
         $tmpWhere = null;
-        if ($sanitarInIDs != "''" && !$this->getCFGDef('ignoreEmpty', '0')) {
+        if ($sanitarInIDs != "''" || $this->getCFGDef('ignoreEmpty', '0')) {
             $tmpWhere = "c.parent IN (" . $sanitarInIDs . ")";
             $tmpWhere .= (($this->getCFGDef('showParent', '0')) ? "" : " AND c.id NOT IN(" . $sanitarInIDs . ")");
         }