+++ assets/snippets/DocLister/snippet.DLvaluelist.php
@@ -13,3 +13,4 @@ foreach ($json as $item) {
     $out[] = $item->pagetitle . '==' . $item->id;
 }
 return implode("||", $out);
+?>